+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 
 - Add [documentation of fragmented list](https://marpit.marp.app/fragmented-list) ([#152](https://github.com/marp-team/marpit/pull/152))
+- Test with Node 12 (Erbium) ([#160](https://github.com/marp-team/marpit/pull/160))
 
 ### Changed