+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add [shorthand for setting text color via image syntax](https://marpit.marp.app/image-syntax?id=shorthand-for-setting-colors) ([#154](https://github.com/marp-team/marpit/issues/154), [#159](https://github.com/marp-team/marpit/pull/159))
 - Add [documentation of fragmented list](https://marpit.marp.app/fragmented-list) ([#152](https://github.com/marp-team/marpit/pull/152))
 - Test with Node 12 (Erbium) ([#160](https://github.com/marp-team/marpit/pull/160))
+- Automate GitHub release ([#161](https://github.com/marp-team/marpit/pull/161))
 
 ### Changed